% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_my_dashboard.R
\name{dashboard_pin_area_percent_time_var}
\alias{dashboard_pin_area_percent_time_var}
\title{Resolve dashboard area percentage time variable}
\usage{
dashboard_pin_area_percent_time_var(time_variable)
}
\description{
Resolve dashboard area percentage time variable
}
\keyword{internal}
