% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_my_dashboard.R
\name{dashboard_pin_data}
\alias{dashboard_pin_data}
\title{Build dashboard pin data}
\usage{
dashboard_pin_data(pin, rv_data, rv_unfiltered = NULL)
}
\description{
Build dashboard pin data
}
\keyword{internal}
