% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_my_dashboard.R
\name{dashboard_pin_params_label}
\alias{dashboard_pin_params_label}
\title{Dashboard pin parameter label}
\usage{
dashboard_pin_params_label(params)
}
\description{
Dashboard pin parameter label
}
\keyword{internal}
