% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_my_dashboard.R
\name{dashboard_pin_plot}
\alias{dashboard_pin_plot}
\title{Build dashboard pin plot}
\usage{
dashboard_pin_plot(pin, rv_data, rv_unfiltered = NULL)
}
\description{
Build dashboard pin plot
}
\keyword{internal}
