% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_my_dashboard.R
\name{dashboard_pin_preview_output}
\alias{dashboard_pin_preview_output}
\title{Dashboard pin preview output}
\usage{
dashboard_pin_preview_output(ns, pin, index)
}
\description{
Dashboard pin preview output
}
\keyword{internal}
