% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_my_dashboard.R
\name{dashboard_pin_table_data}
\alias{dashboard_pin_table_data}
\title{Dashboard pins table data}
\usage{
dashboard_pin_table_data(pin_list, ns = identity)
}
\description{
Dashboard pins table data
}
\keyword{internal}
