% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_my_dashboard.R
\name{dashboard_remove_pin_button}
\alias{dashboard_remove_pin_button}
\title{Dashboard remove button HTML}
\usage{
dashboard_remove_pin_button(index, ns)
}
\description{
Dashboard remove button HTML
}
\keyword{internal}
