% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_global_filter.R, R/fn_prefix_data.R
\name{data_apply_filter}
\alias{data_apply_filter}
\title{Apply a single global filter to a dashboard dataset}
\usage{
data_apply_filter(df, group_var = NULL, group_levels = NULL)

data_apply_filter(df, group_var = NULL, group_levels = NULL)
}
\arguments{
\item{df}{Data frame.}

\item{group_var}{Column name to filter on.}

\item{group_levels}{Values to keep.}
}
\value{
Filtered tibble.

Filtered tibble.
}
\description{
Apply a single global filter to a dashboard dataset

Apply a single global filter to a dashboard dataset
}
