% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_capabilities.R
\name{data_capabilities_check_passed}
\alias{data_capabilities_check_passed}
\title{Read a named validation check}
\usage{
data_capabilities_check_passed(validation_checks, check_name)
}
\arguments{
\item{validation_checks}{Tibble of dataset-level validation checks.}

\item{check_name}{Name of the check to read.}
}
\value{
Logical scalar.
}
\description{
Read a named validation check
}
