% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_capabilities.R
\name{data_capabilities_supports_proportions}
\alias{data_capabilities_supports_proportions}
\title{Determine whether grouped proportions are supported}
\usage{
data_capabilities_supports_proportions(
  validation_checks,
  variable_sets,
  validation_status
)
}
\arguments{
\item{validation_checks}{Dataset-level validation checks.}

\item{variable_sets}{Validated variable inventories.}

\item{validation_status}{Dataset validation status.}
}
\value{
Logical scalar.
}
\description{
Determine whether grouped proportions are supported
}
