% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_density}
\alias{data_density}
\title{Prepare trimmed density-plot data for a numeric variable by group}
\usage{
data_density(data, group_var, num_var, fy_choice)
}
\description{
Prepare trimmed density-plot data for a numeric variable by group
}
