% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_fix_defaults}
\alias{data_fix_defaults}
\title{data_fix_defaults
Restore key factor / date classes after importing a CSV.}
\usage{
data_fix_defaults(df)
}
\arguments{
\item{df}{A data.frame/tibble straight from read_csv()/read.csv()}
}
\value{
Tibble with corrected column types
}
\description{
data_fix_defaults
Restore key factor / date classes after importing a CSV.
}
