% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_group_levels}
\alias{data_group_levels}
\title{Return available levels for a chosen grouping variable}
\usage{
data_group_levels(df, group_var)
}
\arguments{
\item{df}{data-frame containing \code{group_var}.}

\item{group_var}{character(1) - column name.}
}
\value{
character vector of distinct levels, sorted, with \code{NA} removed.
}
\description{
Return available levels for a chosen grouping variable
}
\keyword{internal}
