% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_group_vars}
\alias{data_group_vars}
\title{Return candidate grouping variables from a data-frame}
\usage{
data_group_vars(df, exclude = c("Row_ID", "unique_id"))
}
\arguments{
\item{df}{\code{data.frame} / \code{tibble}}

\item{exclude}{character vector of columns to drop unconditionally.}
}
\value{
character vector of column names in the order they appear in \code{df}.
}
\description{
Factors \emph{or} character columns qualify.  Certain admin columns are excluded.
}
\keyword{internal}
