% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_grouped_dodge_bar}
\alias{data_grouped_dodge_bar}
\title{Summarise a numeric variable by two grouping factors for the most-recent period}
\usage{
data_grouped_dodge_bar(data, time_var, group1_var, group2_var, num_var)
}
\description{
Summarise a numeric variable by two grouping factors for the most-recent period
}
