% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_grouped_lines}
\alias{data_grouped_lines}
\title{Summarise a numeric variable over time, split by a factor, with an "All" line}
\usage{
data_grouped_lines(
  data_filtered,
  data_unfiltered,
  time_var,
  group_var,
  num_var
)
}
\description{
Summarise a numeric variable over time, split by a factor, with an "All" line
}
