% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_line_single_variable}
\alias{data_line_single_variable}
\title{Summarise a single numeric variable over time (filtered & unfiltered)}
\usage{
data_line_single_variable(data_filtered, data_unfiltered, time_var, num_var)
}
\description{
Summarise a single numeric variable over time (filtered & unfiltered)
}
