% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_prop_table_one_group}
\alias{data_prop_table_one_group}
\title{Build a counts-plus-row-percent table for one grouping variable}
\usage{
data_prop_table_one_group(data, var1, var2)
}
\description{
Build a counts-plus-row-percent table for one grouping variable
}
