% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_prop_table_two_groups}
\alias{data_prop_table_two_groups}
\title{Build a counts-plus-row-percent table for two grouping variables}
\usage{
data_prop_table_two_groups(data, var1, var2, var3)
}
\description{
Build a counts-plus-row-percent table for two grouping variables
}
