% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_proportional_area}
\alias{data_proportional_area}
\title{Build proportional-area data over time for a factor}
\usage{
data_proportional_area(data, time_var, group_var)
}
\description{
Build proportional-area data over time for a factor
}
