% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_register_uploads}
\alias{data_register_uploads}
\title{Convert a \code{shiny::fileInput()} data-frame to a named list of datapaths}
\usage{
data_register_uploads(files_df)
}
\arguments{
\item{files_df}{data-frame from \code{input$files}.  Must contain columns \code{name}
and \code{datapath}.}
}
\value{
named list: names are \emph{file stems} (sans extension), values are the
corresponding \code{datapath}s.
}
\description{
Convert a \code{shiny::fileInput()} data-frame to a named list of datapaths
}
\keyword{internal}
