% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_scatter_grouped}
\alias{data_scatter_grouped}
\title{Prepare sampled scatterplot data with grouping}
\usage{
data_scatter_grouped(data, group, x_var, y_var)
}
\description{
Prepare sampled scatterplot data with grouping
}
