% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_simple_bar}
\alias{data_simple_bar}
\title{Summarise one numeric variable by group for the most-recent period}
\usage{
data_simple_bar(data, time_var, group_var, num_var)
}
\description{
Summarise one numeric variable by group for the most-recent period
}
