% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_stacked_area}
\alias{data_stacked_area}
\title{Summarise a numeric variable over time by a grouping factor (area chart data)}
\usage{
data_stacked_area(data, time_var, group_var, num_var)
}
\description{
Summarise a numeric variable over time by a grouping factor (area chart data)
}
