% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_stacked_bar}
\alias{data_stacked_bar}
\title{Summarise a numeric variable by two grouping factors (latest period only)}
\usage{
data_stacked_bar(data, time_var, group1_var, group2_var, num_var)
}
\description{
Summarise a numeric variable by two grouping factors (latest period only)
}
