% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_summary_grouped}
\alias{data_summary_grouped}
\title{Per-period, per-group summary statistics for one numeric variable}
\usage{
data_summary_grouped(data, time, group_var, num_var)
}
\description{
Per-period, per-group summary statistics for one numeric variable
}
