% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_data.R
\name{data_summary_one_variable}
\alias{data_summary_one_variable}
\title{Build per-period summary statistics for one numeric variable}
\usage{
data_summary_one_variable(data, time, var)
}
\description{
Build per-period summary statistics for one numeric variable
}
