% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_add_date_helpers}
\alias{data_user_mapping_add_date_helpers}
\title{Add legacy-compatible date helper fields}
\usage{
data_user_mapping_add_date_helpers(
  data,
  date_column,
  financial_year_start_month = 4L
)
}
\description{
Add legacy-compatible date helper fields
}
