% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_clean_names}
\alias{data_user_mapping_clean_names}
\title{Clean raw column names into simple snake_case}
\usage{
data_user_mapping_clean_names(names)
}
\description{
Clean raw column names into simple snake_case
}
