% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_is_boolean}
\alias{data_user_mapping_is_boolean}
\title{Check whether a selected categorical column is boolean-like}
\usage{
data_user_mapping_is_boolean(x)
}
\description{
Check whether a selected categorical column is boolean-like
}
