% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_match_selected}
\alias{data_user_mapping_match_selected}
\title{Match user selections against original and cleaned names}
\usage{
data_user_mapping_match_selected(selected, column_index)
}
\description{
Match user selections against original and cleaned names
}
