% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_parse_category}
\alias{data_user_mapping_parse_category}
\title{Parse selected categorical columns}
\usage{
data_user_mapping_parse_category(x)
}
\description{
Parse selected categorical columns
}
