% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_parse_number}
\alias{data_user_mapping_parse_number}
\title{Parse selected numeric columns}
\usage{
data_user_mapping_parse_number(x)
}
\description{
Parse selected numeric columns
}
