% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_resolve_columns}
\alias{data_user_mapping_resolve_columns}
\title{Resolve user-selected columns against raw data names}
\usage{
data_user_mapping_resolve_columns(
  data,
  date_column,
  numeric_columns,
  categorical_columns
)
}
\description{
Resolve user-selected columns against raw data names
}
