% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_user_mapping.R
\name{data_user_mapping_standardize}
\alias{data_user_mapping_standardize}
\title{Standardize selected columns into dashboard-ready names and types}
\usage{
data_user_mapping_standardize(data, mapping, financial_year_start_month = 4L)
}
\description{
Standardize selected columns into dashboard-ready names and types
}
