% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validate.R
\name{data_validate_safe_for_summary}
\alias{data_validate_safe_for_summary}
\title{Determine whether a column is safe for numeric summaries}
\usage{
data_validate_safe_for_summary(x, role)
}
\arguments{
\item{x}{Standardized vector.}

\item{role}{Semantic role.}
}
\value{
Logical scalar.
}
\description{
Determine whether a column is safe for numeric summaries
}
