% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_axis_limits.R
\name{mod_axis_limits_server}
\alias{mod_axis_limits_server}
\title{Axis Limits Server}
\usage{
mod_axis_limits_server(id)
}
\arguments{
\item{id}{Module ID.}
}
\value{
Reactive list with \code{min} and \code{max}.
}
\description{
Axis Limits Server
}
\keyword{internal}
