% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_axis_limits.R
\name{mod_axis_limits_ui}
\alias{mod_axis_limits_ui}
\title{Axis Limits UI}
\usage{
mod_axis_limits_ui(
  id,
  title = "Y axis",
  min_label = "Min",
  max_label = "Max",
  reset_label = "Auto"
)
}
\arguments{
\item{id}{Module ID.}

\item{title}{Box title.}

\item{min_label}{Minimum-axis input label.}

\item{max_label}{Maximum-axis input label.}

\item{reset_label}{Reset button label.}
}
\value{
Shiny UI.
}
\description{
Axis Limits UI
}
\keyword{internal}
