% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_actions.R
\name{mod_export_actions_server}
\alias{mod_export_actions_server}
\title{Export Actions Server}
\usage{
mod_export_actions_server(
  id,
  data_r = NULL,
  plot_r = NULL,
  filename_prefix_r = NULL
)
}
\arguments{
\item{id}{Module ID}

\item{data_r}{Reactive returning the data frame for CSV export}

\item{plot_r}{Reactive returning the ggplot object for PNG export}

\item{filename_prefix_r}{Reactive returning the string prefix for the filenames}
}
\description{
Export Actions Server
}
