% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_actions.R
\name{mod_export_actions_ui}
\alias{mod_export_actions_ui}
\title{Export Actions UI}
\usage{
mod_export_actions_ui(id, csv = TRUE, plot = FALSE, pin = FALSE, pin_id = NULL)
}
\arguments{
\item{id}{Module ID}

\item{csv}{Logical, include CSV download button}

\item{plot}{Logical, include Plot download button}

\item{pin}{Logical, include Pin to Dashboard button}

\item{pin_id}{Character, the base namespace ID used for the pinModule (e.g. "plot10_pin").}
}
\description{
Export Actions UI
}
