% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_shared_datatable.R
\name{mod_shared_datatable_ui}
\alias{mod_shared_datatable_ui}
\title{Shared Datatable UI}
\usage{
mod_shared_datatable_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Shared Datatable UI
}
