% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboard_pin.R
\name{pinModuleServer}
\alias{pinModuleServer}
\title{Server logic to capture & store a pinned view}
\usage{
pinModuleServer(id, render_fn, type, params_r, title_r = reactive(type), pins)
}
\arguments{
\item{id}{Module namespace id}

\item{render_fn}{A function \verb{(output, session, out_id) -> NULL} that assigns
the appropriate \verb{render*()} to \code{output[[out_id]]}}

\item{type}{A short character string identifying this view (e.g. "scatter")}

\item{params_r}{A reactive expression returning a named list of inputs}

\item{title_r}{A reactive expression returning a single string title}

\item{pins}{The reactiveVal object storing all pins (pass from parent server)}
}
\value{
Invisibly \code{NULL}; side‐effect is updating a global \code{pins} store
}
\description{
Server logic to capture & store a pinned view
}
