% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dashboard_pin.R
\name{pinModuleUI}
\alias{pinModuleUI}
\title{Wrap Shiny output in a namespaced container (button excluded)}
\usage{
pinModuleUI(id, output_ui_fn)
}
\arguments{
\item{id}{Module namespace id}

\item{output_ui_fn}{A function like \code{plotlyOutput} or \code{DTOutput} (un‐namespaced)
that takes one argument (the output id) and returns UI markup.}
}
\value{
UI output tag (without button)
}
\description{
Wrap Shiny output in a namespaced container (button excluded)
}
