% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_grouped_dodge_bar}
\alias{plot_grouped_dodge_bar}
\title{Build the dodge-bar ggplot for Tab 14}
\usage{
plot_grouped_dodge_bar(
  df,
  group1_var,
  group2_var,
  num_var,
  y_min = NA_real_,
  y_max = NA_real_
)
}
\description{
Build the dodge-bar ggplot for Tab 14
}
