% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_grouped_lines}
\alias{plot_grouped_lines}
\title{Build the ggplot for Tab 11 (grouped line chart)}
\usage{
plot_grouped_lines(
  df,
  time_var,
  group_var,
  num_var,
  time_in,
  y_min = NA_real_,
  y_max = NA_real_
)
}
\description{
Build the ggplot for Tab 11 (grouped line chart)
}
