% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_line_single_variable}
\alias{plot_line_single_variable}
\title{Create the ggplot object for Tab 10}
\usage{
plot_line_single_variable(
  df,
  time_var,
  num_var,
  time_in,
  y_min = NA_real_,
  y_max = NA_real_
)
}
\description{
Create the ggplot object for Tab 10
}
