% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_proportional_area}
\alias{plot_proportional_area}
\title{Build the proportional-area ggplot for Tab 16}
\usage{
plot_proportional_area(df, y_min = NA_real_, y_max = NA_real_)
}
\description{
Build the proportional-area ggplot for Tab 16
}
