% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_scatter_grouped}
\alias{plot_scatter_grouped}
\title{Build grouped scatterplot with linear trend lines}
\usage{
plot_scatter_grouped(
  df,
  group,
  x_var,
  y_var,
  y_min = NA_real_,
  y_max = NA_real_
)
}
\description{
Build grouped scatterplot with linear trend lines
}
