% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_simple_bar}
\alias{plot_simple_bar}
\title{Build the ggplot bar chart for Tab 12}
\usage{
plot_simple_bar(df, group_var, num_var, y_min = NA_real_, y_max = NA_real_)
}
\description{
Build the ggplot bar chart for Tab 12
}
