% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_stacked_area}
\alias{plot_stacked_area}
\title{Build the stacked-area ggplot for Tab 15}
\usage{
plot_stacked_area(
  df,
  time_var,
  group_var,
  num_var,
  y_min = NA_real_,
  y_max = NA_real_
)
}
\description{
Build the stacked-area ggplot for Tab 15
}
