% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_stacked_bar}
\alias{plot_stacked_bar}
\title{Build the stacked-bar ggplot for Tab 13}
\usage{
plot_stacked_bar(
  df,
  group1_var,
  group2_var,
  num_var,
  y_min = NA_real_,
  y_max = NA_real_
)
}
\description{
Build the stacked-bar ggplot for Tab 13
}
