% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{plot_time_breaks}
\alias{plot_time_breaks}
\title{Shared x-axis breaks for time-based plots}
\usage{
plot_time_breaks(df, time_var, time_in)
}
\description{
Shared x-axis breaks for time-based plots
}
\keyword{internal}
