% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{scale_colour_general}
\alias{scale_colour_general}
\title{Shared colour scale for General plots}
\usage{
scale_colour_general(...)
}
\description{
Shared colour scale for General plots
}
\keyword{internal}
