% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_prefix_plot.R
\name{scale_x_time_general}
\alias{scale_x_time_general}
\title{Shared x scale for time-based plots}
\usage{
scale_x_time_general(df, time_var, time_in)
}
\description{
Shared x scale for time-based plots
}
\keyword{internal}
