% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thegeneral.R
\name{the_general_app}
\alias{the_general_app}
\title{Run The General Shiny Application}
\usage{
the_general_app(debug = FALSE)
}
\arguments{
\item{debug}{Enable debug mode with additional logging}
}
\value{
A Shiny application object
}
\description{
Run The General Shiny Application
}
